Moved tags to use swr and client data fetching
