+++ package.json
@@ -27,6 +27,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "slugify": "^1.6.6",
+    "swr": "^2.2.4",
     "typescript": "5.1.6",
     "uuid": "^9.0.1"
   },
